move orderly tests to VersionTests class
