+++ src/app/src/test/kotlin/org/vaccineimpact/orderlyweb/tests/database_tests/VersionTests.kt
@@ -1,8 +1,11 @@
 package org.vaccineimpact.orderlyweb.tests.database_tests
 
+import com.nhaarman.mockito_kotlin.doReturn
+import com.nhaarman.mockito_kotlin.mock
 import org.assertj.core.api.Assertions
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
+import org.vaccineimpact.orderlyweb.ActionContext
 import org.vaccineimpact.orderlyweb.db.Orderly
 import org.vaccineimpact.orderlyweb.db.OrderlyClient
 import org.vaccineimpact.orderlyweb.errors.UnknownObjectError
@@ -246,5 +249,114 @@ class VersionTests : CleanDatabaseTests()
         Assertions.assertThat(results[6].id).isEqualTo("test3versionunpublished")
         Assertions.assertThat(results[6].latestVersion).isEqualTo("test3versionunpublished")
     }
+    @Test
+    fun `getAllReportVersions returns version tags`()
+    {
+        insertReport("report", "v1")
+        insertVersionTags("v1", "c-tag", "a-tag", "b-tag")
 
+        insertReport("report", "v2")
+        insertVersionTags("v2", "aa-tag")
+
+        insertReport("report", "v3")
+
+        val sut = Orderly(isReviewer = true, isGlobalReader = true, reportReadingScopes = listOf())
+        val results = sut.getAllReportVersions()
+
+        assertThat(results[0].id).isEqualTo("v1")
+        //tags should be sorted
+        assertThat(results[0].tags).containsExactlyElementsOf(listOf("a-tag", "b-tag", "c-tag"))
+
+        assertThat(results[1].id).isEqualTo("v2")
+        assertThat(results[1].tags).containsExactlyElementsOf(listOf("aa-tag"))
+
+        assertThat(results[2].id).isEqualTo("v3")
+        assertThat(results[2].tags.count()).isEqualTo(0)
+    }
+
+    @Test
+    fun `getAllReportVersions includes report tags`()
+    {
+        insertReport("report", "v1")
+        insertReportTags("report", "d-tag", "b-tag")
+        insertVersionTags("v1", "c-tag", "a-tag", "b-tag")
+
+        insertReport("report2", "v2")
+        insertVersionTags("v2", "aa-tag")
+
+        insertReport("report3", "v3")
+        insertReportTags("report3", "a-tag")
+
+        val sut = Orderly(isReviewer = true, isGlobalReader = true, reportReadingScopes = listOf())
+        val results = sut.getAllReportVersions()
+
+        assertThat(results[0].id).isEqualTo("v1")
+        assertThat(results[0].tags).containsExactlyElementsOf(listOf("a-tag", "b-tag", "c-tag", "d-tag"))
+
+        assertThat(results[1].id).isEqualTo("v2")
+        assertThat(results[1].tags).containsExactlyElementsOf(listOf("aa-tag"))
+
+        assertThat(results[2].id).isEqualTo("v3")
+        assertThat(results[2].tags).containsExactlyElementsOf(listOf("a-tag"))
+    }
+
+    @Test
+    fun `getAllReportVersions includes orderly tags`()
+    {
+        insertReport("report", "v1")
+        insertVersionTags("v1", "a", "c")
+        insertOrderlyTags("v1", "b", "d")
+
+        insertReport("report2", "v2")
+        insertReportTags("report2", "e")
+        insertOrderlyTags("v2", "f", "e")
+
+        insertReport("report3", "v3")
+        insertOrderlyTags("v3", "g")
+
+        val sut = Orderly(isReviewer = true, isGlobalReader = true, reportReadingScopes = listOf())
+        val results = sut.getAllReportVersions()
+
+        assertThat(results[0].id).isEqualTo("v1")
+        assertThat(results[0].tags).containsExactlyElementsOf(listOf("a", "b", "c", "d"))
+
+        assertThat(results[1].id).isEqualTo("v2")
+        assertThat(results[1].tags).containsExactlyElementsOf(listOf("e", "f"))
+
+        assertThat(results[2].id).isEqualTo("v3")
+        assertThat(results[2].tags).containsExactlyElementsOf(listOf("g"))
+    }
+
+    @Test
+    fun `getAllReportVersions returns report names user is authorized to see`()
+    {
+        insertReport("goodname", "va")
+        insertReport("badname", "vb")
+
+        val mockContext = mock<ActionContext> {
+            on { it.reportReadingScopes } doReturn listOf("goodname")
+        }
+
+        val sut = Orderly(mockContext)
+
+        val result = sut.getAllReportVersions()
+        assertThat(result).hasSize(1)
+        assertThat(result[0].name).isEqualTo("goodname")
+    }
+
+    @Test
+    fun `getAllReportVersions returns all report names if user has global read permissions`()
+    {
+        insertReport("goodname", "va")
+        insertReport("anothername", "vb")
+
+        val mockContext = mock<ActionContext> {
+            on { it.isGlobalReader() } doReturn true
+        }
+
+        val sut = Orderly(mockContext)
+
+        val results = sut.getAllReportVersions()
+        assertThat(results.count()).isEqualTo(2)
+    }
 }
